requirements.txt
apache-beam==2.15.0
apache-flink==1.10.1
avro-python3==1.9.1
certifi==2020.4.5.1
chardet==3.0.4
cloudpickle==1.2.2
crcmod==1.7
dill==0.2.9
docopt==0.6.2
fastavro==0.21.24
future==0.18.2
grpcio==1.28.1
hdfs==2.5.8
httplib2==0.12.0
idna==2.9
mock==2.0.0
numpy==1.18.4
oauth2client==3.0.0
pbr==5.4.5
protobuf==3.11.3
py4j==0.10.8.1
pyarrow==0.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydot==1.4.1
pymongo==3.10.1
pyparsing==2.4.7
python-dateutil==2.8.0
pytz==2020.1
PyYAML==3.13
requests==2.23.0
rsa==4.0
six==1.14.0
urllib3==1.25.9
